Move nextflow.antlr classes into script/config parser packages
