+++ modules/compiler/src/main/antlr/ScriptParser.g4
@@ -41,9 +41,11 @@ options {
 }
 
 @header {
-package nextflow.antlr;
+package nextflow.script.parser;
 
 import org.apache.groovy.parser.antlr4.GroovySyntaxError;
+
+import static nextflow.script.parser.SemanticPredicates.*;
 }
 
 @members {
@@ -356,7 +358,7 @@ assignmentStatement
 expressionStatement
     :   expression
         (
-            { SemanticPredicates.isValidDirective($expression.ctx) }? argumentList
+            { isValidDirective($expression.ctx) }? argumentList
         |
             /* only certain expressions can be called as a directive (no parens) */
         )
