+++ modules/compiler/src/main/java/script/parser/DescriptiveErrorStrategy.java
@@ -16,7 +16,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package nextflow.antlr;
+package nextflow.script.parser;
 
 import org.antlr.v4.runtime.BailErrorStrategy;
 import org.antlr.v4.runtime.CharStream;
